+++ recipes/consul_service.rb
@@ -8,11 +8,9 @@ port = 9090
 consul_name = "prometheus-pod-telemetry"
 
 tag_list = ['prometheus']
-infradb_available = (node.key?('infradb') and node['infradb'].key?('serverInfo') and node['infradb']['serverInfo'].key?('Server') and not node['infradb']['serverInfo']['Server'].nil?)
-location_available = (infradb_available && !node['infradb']['serverInfo']['Server']['Location'].nil?)
+pod = node['rblx_prometheus']['config']['pod']
 
-if infradb_available && location_available
-  pod = node['infradb']['serverInfo']['Server']['Location']['Pod']['Name']
+if pod
   tag_list << pod
   skip_update = node['rblx_prometheus']['skip_consul']
 else