+++ test/integration/default/default_test.rb
@@ -3,18 +3,6 @@
 # The InSpec reference, with examples and extensive documentation, can be
 # found at https://www.inspec.io/docs/reference/resources/
 
-#unless os.windows?
-#  # This is an example test, replace with your own test.
-#  describe user('root'), :skip do
-#    it { should exist }
-#  end
-#end
-#
-## This is an example test, replace it with your own test.
-#describe port(80), :skip do
-#  it { should_not be_listening }
-#end
-
 describe service('docker') do
   it { should be_installed }
   it { should be_enabled }
@@ -41,8 +29,4 @@ end
 
 describe command('curl -XGET -s http://localhost:9090/-/healthy') do
   its('stdout') { should match "Prometheus is Healthy." }
-end
-
-describe command('curl -XGET -s http://localhost:9090/api/v1/targets') do
-  its('stdout') { should match "localhost:9090"}
 end